Обновляющиеся регистры, рабочая кнопка EXECUTE, новый модуль для GUI
